+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/iconpack/newpack/ui/NewPackCreation.kt
@@ -2,6 +2,7 @@ package io.github.composegears.valkyrie.ui.screen.mode.iconpack.newpack.ui
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -10,6 +11,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import io.github.composegears.valkyrie.jewel.settings.CheckboxSettingsRow
+import io.github.composegears.valkyrie.jewel.settings.ExpandedGroup
 import io.github.composegears.valkyrie.jewel.tooling.PreviewTheme
 import io.github.composegears.valkyrie.sdk.compose.foundation.layout.Spacer
 import io.github.composegears.valkyrie.ui.screen.mode.iconpack.common.IconPackEditor
@@ -45,6 +48,18 @@ fun NewIconPackCreation(
             onAddNestedPack = { onAction(NewPackAction.AddNestedPack) },
             onRemoveNestedPack = { onAction(NewPackAction.RemoveNestedPack(it)) },
         )
+        Spacer(32.dp)
+        ExpandedGroup(
+            text = stringResource("iconpack.newpack.creation.additional.options"),
+            paddingValues = PaddingValues(0.dp),
+        ) {
+            CheckboxSettingsRow(
+                text = stringResource("iconpack.newpack.creation.use.material.pack"),
+                infoText = stringResource("iconpack.newpack.creation.use.material.pack.description"),
+                checked = state.useMaterialPack,
+                onCheckedChange = { onAction(NewPackAction.UseMaterialPack(it)) },
+            )
+        }
         Spacer(32.dp)
         Row(
             modifier = Modifier.fillMaxWidth(),
@@ -84,6 +99,7 @@ private fun NewIconPackCreationPreview() = PreviewTheme(alignment = Alignment.To
                     ),
                 ),
             ),
+            useMaterialPack = true,
         ),
         onAction = {},
         onValueChange = {},